Restore inline observation header
